backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
requests==2.31.0
sqlparse==0.5.0
stripe==9.6.0
typing_extensions==4.11.0
urllib3==2.2.1
